requirements.txt
langchain
openai
langGraph
langchain-community
langchain-groq
langchain-openai
langchain-core
faiss-cpu
tiktoken
python-dotenv
streamlit
